upstream(core): move `ConsensusCommitOutput` into its own module
